+++ src/main/java/com/webapp/springBoot/service/PostUsersAppService.java
@@ -1,5 +1,6 @@
 package com.webapp.springBoot.service;
 
+import com.webapp.springBoot.DTO.Kafka.RequestFeedDTO;
 import com.webapp.springBoot.DTO.Post.RequestPostDTO;
 import com.webapp.springBoot.DTO.Post.ResponseListPostDTO;
 import com.webapp.springBoot.DTO.Post.ResponsePostDTO;
@@ -9,10 +10,14 @@ import com.webapp.springBoot.entity.UsersApp;
 import com.webapp.springBoot.exception.validation.ValidationErrorWithMethod;
 import com.webapp.springBoot.repository.PostsUsersAppRepository;
 import com.webapp.springBoot.repository.UsersAppRepository;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.kafka.clients.producer.RecordMetadata;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.cache.annotation.Caching;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.kafka.support.SendResult;
 import org.springframework.transaction.annotation.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -36,7 +41,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
 
+@Slf4j
 @Service
 public class PostUsersAppService {
     @Autowired
@@ -51,14 +58,21 @@ public class PostUsersAppService {
     @Autowired
     private PostsUsersAppRepository postsUsersAppRepository;
 
+    @Autowired
+    private KafkaTemplate<String, String> kafkaTemplate;
+
     // <------------------------ ПРОВЕРКА В СУЩНОСТИ PostUsersAppService-------------------------->
     public void checkPostUserByNicknameUser(UsersApp usersApp, String nicknameUser){
         if(!usersApp.equals(usersService.findUsersByNickname(nicknameUser))){
             throw new LockedException("Пользователь не имеет право управлять данным постом");
         }
     }
-    public PostsUserApp checkPostUserByNicknameUser(String postName, String nicknameUser){
-        PostsUserApp postsUserApp = findByName(postName);
+    public PostsUserApp checkPostUserByNicknameUser(String namePost, String nicknameUser){
+        Optional<PostsUserApp> optionalPostsUserApp = findByName(namePost);
+        if (optionalPostsUserApp.isEmpty()) {
+            throw new NoSuchElementException("Пост не найден");
+        }
+        PostsUserApp postsUserApp = optionalPostsUserApp.get();
         if(postsUserApp.getUsersApp().equals(usersService.findUsersByNickname(nicknameUser))){
             return postsUserApp;
         } else {
@@ -108,7 +122,11 @@ public class PostUsersAppService {
     }
      @Cacheable(value = "POST", key = "#namePost")
      public ResponsePostDTO getPost(String namePost){
-        PostsUserApp postsUserApp = findByName(namePost);
+         Optional<PostsUserApp> optionalPostsUserApp = findByName(namePost);
+         if (optionalPostsUserApp.isEmpty()) {
+             throw new NoSuchElementException("Пост не найден");
+         }
+         PostsUserApp postsUserApp = optionalPostsUserApp.get();
          boolean set;
          LocalDateTime localDateTime;
          if(postsUserApp.getUpdateDate() != null){
@@ -121,13 +139,27 @@ public class PostUsersAppService {
         return new ResponsePostDTO(postsUserApp.getTitle(), postsUserApp.getDescription(),postsUserApp.getName(), postsUserApp.getUsersApp().getNickname() , filePostsUsersAppService.getFileName(postsUserApp), localDateTime, set);
      }
 
-    // <------------------------ ПОИСК В СУЩНОСТИ PostUsersAppService-------------------------->
-    public PostsUserApp findByName(String name) {
-        Optional<PostsUserApp> optionalPostsUserApp = postsUsersAppRepository.findByName(name);
+    public ResponsePostDTO getPostNull(String namePost){
+        Optional<PostsUserApp> optionalPostsUserApp = findByName(namePost);
         if (optionalPostsUserApp.isEmpty()) {
-            throw new NoSuchElementException("Пост не найден");
+            return null;
         }
-        return optionalPostsUserApp.get();
+        PostsUserApp postsUserApp = optionalPostsUserApp.get();
+        boolean set;
+        LocalDateTime localDateTime;
+        if(postsUserApp.getUpdateDate() != null){
+            set = true;
+            localDateTime = postsUserApp.getUpdateDate();
+        } else {
+            set = false;
+            localDateTime = postsUserApp.getCreateDate();
+        }
+        return new ResponsePostDTO(postsUserApp.getTitle(), postsUserApp.getDescription(),postsUserApp.getName(), postsUserApp.getUsersApp().getNickname() , filePostsUsersAppService.getFileName(postsUserApp), localDateTime, set);
+    }
+
+    // <------------------------ ПОИСК В СУЩНОСТИ PostUsersAppService-------------------------->
+    public Optional<PostsUserApp>  findByName(String name) {
+        return postsUsersAppRepository.findByName(name);
     }
 
 
@@ -139,7 +171,11 @@ public class PostUsersAppService {
     })
     @Transactional
     public void deletePostUsersApp(String namePost, String nicknameUser) throws IOException {
-        PostsUserApp postsUserApp = findByName(namePost);
+        Optional<PostsUserApp> optionalPostsUserApp = findByName(namePost);
+        if (optionalPostsUserApp.isEmpty()) {
+            throw new NoSuchElementException("Пост не найден");
+        }
+        PostsUserApp postsUserApp = optionalPostsUserApp.get();
         checkPostUserByNicknameUser(postsUserApp.getUsersApp(), nicknameUser);
         postsUserApp.setUsersApp(null);
         filePostsUsersAppService.deleteFileTapeUsersAppService(postsUserApp);
@@ -175,6 +211,16 @@ public class PostUsersAppService {
         }
         postsUserApp.generateName();
         postsUsersAppRepository.save(postsUserApp);
+        CompletableFuture<SendResult<String, String>> future =  kafkaTemplate.send("news-feed-topic-user", nicknameUser, postsUserApp.getName());
+        future.whenComplete((result, exception) -> {
+            if(exception != null){
+                log.error("Возникла ошибка добавления поста в ленту", exception);
+            } else {
+                RecordMetadata meta = result.getRecordMetadata();
+                log.info("Сообщение отправлено в топик '{}' партиция {} офсэт {}",
+                        meta.topic(), meta.partition(), meta.offset());
+            }
+        });
         return new ResponsePostDTO(postsUserApp.getTitle(), postsUserApp.getDescription(),postsUserApp.getName(), postsUserApp.getUsersApp().getNickname() , filePostsUsersAppService.getFileName(postsUserApp), LocalDateTime.now(), false);
 
     }
